detect/requirements.txt
essentia-tensorflow>=2.1
fastapi>=0.104.0
uvicorn>=0.24.0
librosa>=0.10.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
python-multipart>=0.0.6
requests>=2.31.0